Change br-lan control
Jira-ID: SECO-8033, SECO-8006

Bridge control changed so that br-lan is not deleted when
CBMA/MDM Agent is stopped. Also trying to use static MAC
for the bridge using MAC from some physical interfaces
(osf or ethernet) that is not configured to lower CBMA.
Default configs changed so that eth0/end0/end or lan1 interface is addde to bridge.

Signed-off-by: Saku Auvinen <saku.auvinen@tii.ae>